backend/requirements.txt
fastapi
uvicorn
pydantic
joblib
transformers
google-api-python-client
google-auth
requests